# src/Talabat/Talabat.API/openapi/customer-api.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Talabat.Customer.API | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/me/orders": {
      "get": {
        "tags": [
          "Orders"
        ],
        "operationId": "GetOrders",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "oauth2": [
              "customer.api"
            ]
          }
        ]
      }
    },
    "/api/me/orders/{orderId}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "operationId": "GetOrderDetails",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetailResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "oauth2": [
              "customer.api"
            ]
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "MoneyDto": {
        "required": [
          "amount"
        ],
        "type": "object",
        "properties": {
          "amount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "currency": {
            "type": "string",
            "default": "EGP"
          }
        }
      },
      "OrderDeliveryAddressDto": {
        "required": [
          "street",
          "city",
          "buildingNumber",
          "floor"
        ],
        "type": "object",
        "properties": {
          "street": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "buildingNumber": {
            "type": "string"
          },
          "floor": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "OrderDetailResponse": {
        "required": [
          "id",
          "customerId",
          "restaurantId",
          "deliveryAddress",
          "items",
          "total",
          "placedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "customerId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "restaurantId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "deliveryAddress": {
            "$ref": "#/components/schemas/OrderDeliveryAddressDto"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderLineItemDto"
            }
          },
          "total": {
            "$ref": "#/components/schemas/MoneyDto"
          },
          "placedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "OrderLineItemDto": {
        "required": [
          "productName",
          "unitPrice",
          "quantity",
          "lineTotal"
        ],
        "type": "object",
        "properties": {
          "productName": {
            "type": "string"
          },
          "unitPrice": {
            "$ref": "#/components/schemas/MoneyDto"
          },
          "quantity": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "lineTotal": {
            "$ref": "#/components/schemas/MoneyDto"
          }
        }
      },
      "OrderListResponse": {
        "required": [
          "items",
          "page",
          "pageSize",
          "totalCount"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderSummaryDto"
            }
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "OrderSummaryDto": {
        "required": [
          "id",
          "restaurantId",
          "total",
          "placedAt",
          "itemCount"
        ],
        "type": "object",
        "properties": {
          "id": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "restaurantId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "total": {
            "$ref": "#/components/schemas/MoneyDto"
          },
          "placedAt": {
            "type": "string",
            "format": "date-time"
          },
          "itemCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "oauth2": {
        "type": "oauth2",
        "description": "OIDC authorization code flow with PKCE against Talabat.Identity.",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://localhost:7237/connect/authorize",
            "tokenUrl": "https://localhost:7237/connect/token",
            "scopes": {
              "openid": "Subject identifier",
              "profile": "Profile claims",
              "roles": "Role claims",
              "customer.api": "Customer API access"
            }
          }
        }
      }
    }
  },
...
}